bug(fix) create connectorPortService (EWC-374)
