+++ src/services/connector-service.js
@@ -17,13 +17,7 @@ const ConnectorManager = require('../sequelize/managers/connector-manager');
 const Errors = require('../helpers/errors');
 const ErrorMessages = require('../helpers/error-messages');
 const AppHelper = require('../helpers/app-helper');
-const https = require('https');
-const http = require('http');
-const constants = require('../helpers/constants');
-const logger = require('../logger');
-const qs = require('qs');
 const Op = require('sequelize').Op;
-const fs = require('fs');
 const ConnectorPortManager = require('../sequelize/managers/connector-port-manager');
 const MicroserviceService = require('../services/microservices-service');
 
@@ -81,92 +75,6 @@ async function getConnectorList(transaction) {
   return await ConnectorManager.findAll({}, transaction)
 }
 
-async function openPortOnRandomConnector(isPublicAccess, transaction) {
-  let isConnectorPortOpen = false;
-  let ports = null;
-  let connector = null;
-  const maxAttempts = 5;
-  for (let i = 0; i < maxAttempts; i++) {
-    try {
-      connector = await _getRandomConnector(transaction);
-      ports = await _openPortsOnConnector(connector, isPublicAccess);
-      if (ports) {
-        isConnectorPortOpen = true;
-        break;
-      }
-    } catch (e) {
-      logger.warn(`Failed to open ports on Connector. Attempts ${i + 1}/${maxAttempts}`)
-    }
-  }
-  if (!isConnectorPortOpen) {
-    throw new Error('Not able to open port on remote Connector. Gave up after 5 attempts.')
-  }
-  ports.connectorId = connector.id;
-  return {ports: ports, connector: connector}
-}
-
-async function _makeRequest(connector, options, data) {
-  return new Promise((resolve, reject) => {
-    let httpreq = (connector.devMode ? http : https).request(options, function (response) {
-      console.log(response.statusCode);
-      let output = '';
-      response.setEncoding('utf8');
-
-      response.on('data', function (chunk) {
-        output += chunk;
-      });
-
-      response.on('end', function () {
-        let responseObj = JSON.parse(output);
-        console.log(responseObj);
-        if (responseObj.errormessage) {
-          return reject(new Error(responseObj.errormessage));
-        } else {
-          return resolve(responseObj);
-        }
-      });
-    });
-
-    httpreq.on('error', function (err) {
-      console.log(err);
-      if (err instanceof Error)
-        return reject(new Error(err.message));
-      else
-        return reject(new Error(JSON.stringify(err)));
-    });
-
-    httpreq.write(data);
-    httpreq.end();
-  })
-}
-
-async function closePortOnConnector(connector, ports) {
-  let data = qs.stringify({
-    mappingid: ports.mappingId
-  });
-  console.log(data);
-
-  let port = connector.devMode ? constants.CONNECTOR_HTTP_PORT : constants.CONNECTOR_HTTPS_PORT;
-
-  let options = {
-    host: connector.domain,
-    port: port,
-    path: '/api/v2/mapping/remove',
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      'Content-Length': Buffer.byteLength(data)
-    }
-  };
-  if (!connector.devMode && connector.cert && connector.isSelfSignedCert === true) {
-    const ca = fs.readFileSync(connector.cert);
-    options.ca = new Buffer.from(ca);
-  }
-
-
-  await _makeRequest(connector, options, data)
-}
-
 function _validateConnectorData(connectorData) {
   if (connectorData.domain) {
     const validDomain = AppHelper.isValidDomain(connectorData.domain) || AppHelper.isValidPublicIP(connectorData.domain);
@@ -180,53 +88,9 @@ function _validateConnectorData(connectorData) {
   }
 }
 
-async function _openPortsOnConnector(connector, isPublicAccess) {
-  let data = isPublicAccess
-    ? await qs.stringify({
-      mapping: '{"type":"public","maxconnections":60,"heartbeatabsencethreshold":200000}'
-    })
-    : await qs.stringify({
-      mapping: '{"type":"private","maxconnectionsport1":1, "maxconnectionsport2":1, ' +
-      '"heartbeatabsencethresholdport1":200000, "heartbeatabsencethresholdport2":200000}'
-    });
-
-  let port = connector.devMode ? constants.CONNECTOR_HTTP_PORT : constants.CONNECTOR_HTTPS_PORT;
-
-  let options = {
-    host: connector.domain,
-    port: port,
-    path: '/api/v2/mapping/add',
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      'Content-Length': Buffer.byteLength(data)
-    }
-  };
-  if (!connector.devMode && connector.cert && connector.isSelfSignedCert === true) {
-    const ca = fs.readFileSync(connector.cert);
-    options.ca = new Buffer.from(ca);
-  }
-
-  const ports = await _makeRequest(connector, options, data);
-  return ports
-}
-
-async function _getRandomConnector(transaction) {
-  const connectors = await ConnectorManager.findAll({}, transaction);
-
-  if (connectors && connectors.length > 0) {
-    const randomNumber = Math.round((Math.random() * (connectors.length - 1)));
-    return connectors[randomNumber]
-  } else {
-    throw new Error('no connectors defined')
-  }
-}
-
 module.exports = {
   createConnector: TransactionDecorator.generateTransaction(createConnector),
   updateConnector: TransactionDecorator.generateTransaction(updateConnector),
   deleteConnector: TransactionDecorator.generateTransaction(deleteConnector),
-  getConnectorList: TransactionDecorator.generateTransaction(getConnectorList),
-  openPortOnRandomConnector: openPortOnRandomConnector,
-  closePortOnConnector: closePortOnConnector
+  getConnectorList: TransactionDecorator.generateTransaction(getConnectorList)
 };
